Added passthrough for page render calls. (#16)
This update ships an improved Serializable#render method which checks
for page render requests and if such a request is present it just passes
the options down to Rails. Otherwise the regular serializer logic
applies and the fast_jsonapi options helper is called.

With this fix in place we allow applications to mix API actions and
regular page view actions. Thats quite handy in combination with other
gems like administrate, instrumentation frameworks, or
monitoring/profiling solutions.

Signed-off-by: Hermann Mayer <hermann.mayer92@gmail.com>